feat: allow operatorSets to be registered from context (#253)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
As an AVS developer, I want to create and allocate an OperatorSet and
later use it in my project without having it defined at setup

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- Combines results from `getOperatorSets` script with `OperatorSets`
defined in the context

**Result:**
<!--
*After your change, what will change.
-->
- Registers template and context defined `OperatorSets`

**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
- Extends tests to cover new yaml helpers

**Open questions:**
<!--
(optional) Any open questions or feedback on design desired?
-->
- Should we expose a new command `devkit avs register-operators` to
allow changes in `.hourglass/context/*.yaml` to be processed against the
context defined `OperatorSets` when they come in to use or is this to be
handled by opCtl? cc. @bdchatham