feat: integrate ECDSA registration for KeyRegistrar (#269)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
As a devkit user, I want to be able to register either BN254 or ECDSA
keys for my Operator to the KeyRegsitrar

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- Allows either curveType to be registered to the KeyRegistrar

**Result:**
<!--
*After your change, what will change.
-->
- Can setup and register ECDSA or BN254 keys in the
`RegisterKeyInKeyRegistrarAction` deploy step