chore: rm registrar address from wizard (#278)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
The Registrar address will not be available on context creation and is
an output after the devkit workflow. devkit currently isn't factored to
manage existing registar addresses given devkit is used to deploy it.
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
**Result:**
<!--
*After your change, what will change.
-->
**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
**Open questions:**
<!--
(optional) Any open questions or feedback on design desired?
-->