fix: create tmp directory in working directory (#287)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
As a devkit user that wants to upgrade to the latest template changes, I
want any temp directory that is created to hold the correct permissions
so that I do not need any `$TMPDIR` workarounds.

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- Creates upgrade temp directory in cwd under `temp_internal`
- Checks working tree is clean before upgrading
- Adds `temp_internal` entry to `.gitignore` if missing and commits
changes

**Result:**
<!--
*After your change, what will change.
-->
- Successful upgrades within the cwd

**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
- Unit tests pass and manually tested against VRF

**Open questions:**
<!--
(optional) Any open questions or feedback on design desired?
-->
- ~~We should perform an upgrade to ensure the temp directory we use is
present in the projects `.gitignore`. If the `.gitignore` does not have
the appropriate entry, upgrades will fail with `Uncommitted changes
found`~~