fix: ensure migrations are context aware (#297)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
As a devkit user upgrading to the most recent version of devkit, I want
any addresses that are updated in my context to be updated according to
which network the context represents.

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- All `mainnet` and `sepolia` addresses are recorded as constants and
are selected according to the networks `chainId`

**Result:**
<!--
*After your change, what will change.
-->
- `mainnet` contexts are correctly updated with `mainnet` values

**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
- Tested in all environments

**Open questions:**
<!--
(optional) Any open questions or feedback on design desired?
-->
- Could zeus be extended to pull middleware addresses removing the need
for this context aware upgrade?