fix: skip ensure docker for L1 deploy outside of devnet (#294)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
As a devkit user deploying to testnet/mainnet, I do not need to have
docker running and the deployment should not start docker for me.

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- Moves `EnsureDockerIsRunning` call to devnet branches

**Result:**
<!--
*After your change, what will change.
-->
- Docker is only ensured for devnet deployments